+++ scripts/deb/ubuntu-focal/deps.sh
@@ -8,10 +8,10 @@ export LD_LIBRARY_PATH=$CMAKE_PREFIX_PATH/lib
 
 wget -qO- https://curl.se/download/curl-8.7.1.tar.xz | tar Jxf - -C /tmp
 wget -qO- https://downloads.videolan.org/pub/videolan/dav1d/1.4.1/dav1d-1.4.1.tar.xz | tar Jxf - -C /tmp
-wget -qO- https://ffmpeg.org/releases/ffmpeg-6.1.1.tar.xz | tar Jxf - -C /tmp
+wget -qO- https://ffmpeg.org/releases/ffmpeg-7.0.tar.xz | tar Jxf - -C /tmp
 wget -qO- https://github.com/mpv-player/mpv/archive/v0.36.0.tar.gz | tar zxf - -C /tmp
 git clone https://github.com/dragonflylee/glfw.git -b switchfin --depth=1 /tmp/glfw
-git clone https://git.videolan.org/git/ffmpeg/nv-codec-headers.git -b n12.1.14.0 --depth=1 /tmp/nv-codec-headers
+git clone https://git.videolan.org/git/ffmpeg/nv-codec-headers.git -b n12.2.72.0 --depth=1 /tmp/nv-codec-headers
 
 cd /tmp/curl-8.7.1
 cmake -B build -G Ninja -DCMAKE_BUILD_TYPE=Release -DCMAKE_INSTALL_PREFIX=$CMAKE_PREFIX_PATH \
@@ -29,7 +29,7 @@ meson install -C build
 
 make PREFIX=$CMAKE_PREFIX_PATH -C /tmp/nv-codec-headers install
 
-cd /tmp/ffmpeg-6.1.1
+cd /tmp/ffmpeg-7.0
 ./configure --prefix=$CMAKE_PREFIX_PATH --enable-shared --disable-static \
   --ld=g++ --enable-nonfree --enable-openssl --enable-libv4l2 \
   --enable-opengl --enable-libass --disable-doc --enable-asm --enable-rpath \
@@ -40,6 +40,7 @@ make -j$(nproc)
 make install
 
 cd /tmp/mpv-0.36.0
+patch -Nbp1 -i /opt/scripts/mingw64/mpv/0002-lavfi-channel-layout.patch
 meson setup build --prefix=$CMAKE_PREFIX_PATH --libdir=lib --buildtype=release --default-library=shared \
   -Dlibmpv=true -Dcplayer=false -Dtests=false -Dlibarchive=disabled -Dlua=disabled
 meson compile -C build